Train poisson regresion models

Trains poisson regression models

Usage

trainPoissonModels(
  connectionDetails,
  cdmDatabaseSchema,
  exposureDatabaseSchema,
  exposureTable,
  outcomeDatabaseSchema,
  outcomeTable,
  outcomeIds,
  covariateSettings,
  saveDir,
  ...
)

Arguments

connectionDetails

The connection details to the database. Should be an object of type createConnectionDetails.

cdmDatabaseSchema

The database schema where the cdm database is stored.

exposureDatabaseSchema

The database schema where table with the exposure cohorts are stored.

exposureTable

The table where the exposure cohorts are stored.

outcomeDatabaseSchema

The database schema where table with the outcome cohorts are stored.

outcomeTable

The table where the outcome cohorts are stored.

outcomeIds

The cohort definition ids of the outcome cohorts.

covariateSettings

The covariates that will be used for fitting the poisson regression models. An object of type covariateSettings, generated with createCovariateSettings.

saveDir

The directory where the models will be stored

...

Additional parameters passed to fitCyclopsModel