Generate cohort observation period

Generates new cohort observation periods for the patients in the exposure cohorts. At the moment, the new observation period is created by limiting the actual observation period at the end of the exposure for each patient.

Usage

generateCohortObservationPeriod(
  connectionDetails,
  cdmDatabaseSchema,
  exposureDatabaseSchema,
  exposureTable,
  resultDatabaseSchema,
  resultTable = "cohort_observation_period",
  maxObservationPeriod,
  washoutPeriod
)

Arguments

connectionDetails

The connection details to the database. Should be an object of type createConnectionDetails.

cdmDatabaseSchema

The database schema where the cdm database is stored.

exposureDatabaseSchema

The database schema where table with the exposure cohorts are stored.

exposureTable

The table where the exposure cohorts are stored.

resultDatabaseSchema

The database schema where the table with the new observation periods will be stored. Needs write permission.

resultTable

The table with the new obsrevation periods.

maxObservationPeriod

The maximum length of the observation period.

washoutPeriod

The minimum number of days before exposure.